feat: transfer album database to postgress
