+++ src/album/album.module.ts
@@ -1,20 +1,19 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { AlbumService } from './album.service';
 import { AlbumController } from './album.controller';
-import { AlbumRepository } from './album.repository';
 import { TrackModule } from '../track/track.module';
 import { FavsModule } from '../favs/favs.module';
-// import { AlbumEntity } from './entities/album.entity';
-// import { TypeOrmModule } from '@nestjs/typeorm';
+import { AlbumEntity } from './entities/album.entity';
+import { TypeOrmModule } from '@nestjs/typeorm';
 
 @Module({
   imports: [
-    // TypeOrmModule.forFeature([AlbumEntity]),
+    TypeOrmModule.forFeature([AlbumEntity]),
     forwardRef(() => TrackModule),
     forwardRef(() => FavsModule),
   ],
   controllers: [AlbumController],
-  providers: [AlbumService, AlbumRepository],
+  providers: [AlbumService],
   exports: [AlbumService],
 })
 export class AlbumModule {}
