+++ src/album/album.service.ts
@@ -1,14 +1,20 @@
 import { Injectable, Inject, forwardRef } from '@nestjs/common';
 import { CreateAlbumDto } from './dto/create-album.dto';
 import { UpdateAlbumDto } from './dto/update-album.dto';
-import { AlbumRepository } from './album.repository';
 import { TrackService } from '../track/track.service';
 import { FavsService } from '../favs/favs.service';
+import { AlbumEntity } from './entities/album.entity';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
+import { Album } from './album.types';
+import { v4 as uuidv4 } from 'uuid';
+import { NotFoundException } from '../lib/exception';
 
 @Injectable()
 export class AlbumService {
   constructor(
-    private readonly albumRepository: AlbumRepository,
+    @InjectRepository(AlbumEntity)
+    private readonly albumRepository: Repository<AlbumEntity>,
     @Inject(forwardRef(() => TrackService))
     private readonly trackService: TrackService,
     @Inject(forwardRef(() => FavsService))
@@ -16,27 +22,46 @@ export class AlbumService {
   ) {}
 
   async create(createAlbumDto: CreateAlbumDto) {
-    return await this.albumRepository.create(createAlbumDto);
+    const album: Album = {
+      id: uuidv4(),
+      ...createAlbumDto,
+    };
+
+    return await this.albumRepository.save(album);
   }
 
   async findAll() {
-    return await this.albumRepository.findAll();
+    return await this.albumRepository.find();
   }
 
   async findOne(id: string) {
-    return await this.albumRepository.findOne(id);
+    const album = await this.albumRepository.findOne({ where: { id: id } });
+    if (!album) throw new NotFoundException(id);
+    return album;
   }
 
   async update(id: string, updateAlbumDto: UpdateAlbumDto) {
-    return await this.albumRepository.update(id, updateAlbumDto);
+    const album = await this.findOne(id);
+    Object.assign(album, updateAlbumDto);
+    return await this.albumRepository.save(album);
   }
 
   async remove(id: string) {
-    await this.albumRepository.remove(id);
+    const artist = await this.findOne(id);
+    await this.albumRepository.remove(artist);
     this.trackService.updateAlbumId(id);
   }
 
   async updateArtistId(artistId: string) {
-    this.albumRepository.updateArtistId(artistId);
+    let albums = await this.albumRepository.find();
+
+    albums = albums.map((album) => ({
+      ...album,
+      artistId: album.artistId === artistId ? null : album.artistId,
+    }));
+
+    for (let album of albums) {
+      await this.albumRepository.save(album);
+    }
   }
 }
