+++ src/artist/artist.module.ts
@@ -1,7 +1,6 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { ArtistService } from './artist.service';
 import { ArtistController } from './artist.controller';
-import { ArtistRepository } from './artist.repository';
 import { TrackModule } from '../track/track.module';
 import { AlbumModule } from 'src/album/album.module';
 import { FavsModule } from 'src/favs/favs.module';
@@ -16,7 +15,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
     forwardRef(() => FavsModule),
   ],
   controllers: [ArtistController],
-  providers: [ArtistService, ArtistRepository],
+  providers: [ArtistService],
   exports: [ArtistService],
 })
 export class ArtistModule {}
